requirements.txt
streamlit==1.36.0
numpy==1.26.4
keras==3.3.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
tensorflow==2.16.1
pillow==10.3.0
pickleshare==0.7.5
